added support for 480p.  Updated mediasite.rss() to use download().  tgx bug fix.

+++ delugeonal/mediasite.py
@@ -36,6 +36,18 @@ class site(ABC):
         pass
 
     def download(self, downloads, args = minorimpact.default_arg_flags):
+        """Download torrent files.
+
+        Parameters
+        ----------
+        list
+            tuple
+                str
+                    Name of the torrent file.
+                str
+                    Download url.
+        """
+
         if (downloads is None):
             raise Exception(f"downloads is not defined")
 
@@ -45,12 +57,12 @@ class site(ABC):
             if ('codec' not in parsed):
                 if (args.debug): print(f"couldn't parse codec from {name}")
                 continue
-            if ('resolution' not in parsed):
-                if (args.debug): print(f"couldn't parse resolution from {name}")
-                continue
             if ('title' not in parsed or 'season' not in parsed or 'episode' not in parsed):
                 if (args.debug): print(f"couldn't parse title, season or episode from {name}")
                 continue
+            if ('resolution' not in parsed):
+                # I have reservations about this.
+                parsed['resolution'] = '480p'
             if (args.debug): print(f"{parsed}")
 
             item = { 'name':name, 'title':parsed['title'], 'season':parsed['season'], 'episode':parsed['episode'], 'url':url, 'codec':parsed['codec'], 'resolution':parsed['resolution']}
@@ -79,6 +91,8 @@ class site(ABC):
                 continue
             uravo.event({'AlertGroup':'db_title', 'AlertKey':item_title, 'Severity':'green', 'Summary':f"Can't get {db.name} title for {item_title}"})
 
+            # Apply transformations to the "official" name.  This is where the user gets to override the wisdom of the masses for their own
+            #   nefarious ends.
             transformation = delugeonal.transform(title, item['season'], item['episode'])
             if (transformation is not None):
                 if (args.verbose): print(f" ... applying transformation: '{title}'=>'{transformation['title']}', season {item['season']}=>{transformation['season']}, episode {item['episode']}=>{transformation['episode']}")
@@ -112,13 +126,15 @@ class site(ABC):
             if (os.path.exists(config['default']['download_dir']) is False):
                 raise Exception(f"{config['default']['download_dir']} does not exist.")
             torrent_filename = re.sub(" ", ".", item['name']) + ".torrent"
-            if args.verbose: print(f" ... downloading {torrent_filename} to {config['default']['download_dir']}")
-            if (args.dryrun is False):
-                r = requests.get(link_url, stream=True)
-                with open(f"{config['default']['download_dir']}/{torrent_filename}", "wb") as f:
-                   for chunk in r.iter_content(chunk_size=128):
-                      f.write(chunk)
-                self.cache['downloads'][episode_key]['date'] = datetime.now()
+            c = 'y' if (args.yes) else minorimpact.getChar(default='y', end='\n', prompt=f"Download {torrent_filename} to {config['default']['download_dir']}? (Y/n) ", echo=True).lower()
+            if (c == 'y'):
+                if args.verbose: print(f" ... downloading {torrent_filename} to {config['default']['download_dir']}")
+                if (args.dryrun is False):
+                    r = requests.get(link_url, stream=True)
+                    with open(f"{config['default']['download_dir']}/{torrent_filename}", "wb") as f:
+                       for chunk in r.iter_content(chunk_size=128):
+                          f.write(chunk)
+                    self.cache['downloads'][episode_key]['date'] = datetime.now()
         
     def fill(self, search_string, args=minorimpact.default_arg_flags):
         show = server.show_name(search_string)
@@ -172,85 +188,7 @@ class site(ABC):
         if (self.rss_url is None):
             raise Exception("rss_url is not defined")
 
-        for name,url in self.rss_feed():
-            parsed = PTN.parse(name)
-                
-            if ('codec' not in parsed):
-                if (args.debug): print(f"couldn't parse codec from {name}")
-                continue
-            if ('resolution' not in parsed):
-                if (args.debug): print(f"couldn't parse resolution from {name}")
-                continue
-            if ('title' not in parsed or 'season' not in parsed or 'episode' not in parsed):
-                if (args.debug): print(f"couldn't parse title, season or episode from {name}")
-                continue
-            if (args.debug): print(f"{parsed}")
-
-            item = { 'name':name, 'title':parsed['title'], 'season':parsed['season'], 'episode':parsed['episode'], 'url':url, 'codec':parsed['codec'], 'resolution':parsed['resolution']}
-            if(args.debug): print(f"{item}")
-
-            item_title = f"{item['title']} ({item['year']})" if 'year' in item else item['title']
-            if (args.debug): print(f"{item_title}")
-            if (item['codec'] != config['default']['codec']):
-                if (args.debug): print(f"invalid codec ({item['codec']}!={config['default']['codec']})")
-                continue
-            if (item['resolution'] != config['default']['resolution']):
-                if (args.debug): print(f"invalid resolution ({item['resolution']}!={config['default']['resolution']})")
-                continue
-
-            if (args.verbose): print(f"processing {item_title} S{item['season']}E{item['episode']}:")
-            title = None
-            for db in mediadbs:
-                if (db.istype('tv')):
-                    title = db.get_title(item_title, year = True, headless=args.yes)
-                    if (title is not None):
-                        break
-            if (title is None):
-                uravo.event({'AlertGroup':'db_title', 'AlertKey':item_title, 'Severity':'yellow', 'Summary':f"Can't get {db.name} title for {item_title}"})
-                if (args.verbose): print(f" ... FAILED: couldn't find {db.name} title for {item_title}")
-                continue
-            uravo.event({'AlertGroup':'db_title', 'AlertKey':item_title, 'Severity':'green', 'Summary':f"Can't get {db.name} title for {item_title}"})
-
-            transformation = delugeonal.transform(title, item['season'], item['episode'])
-            if (transformation is not None):
-                if (args.verbose): print(f" ... applying transformation: '{title}'=>'{transformation['title']}', season {item['season']}=>{transformation['season']}, episode {item['episode']}=>{transformation['episode']}")
-                title = transformation['title']
-                item['season'] = transformation['season']
-                item['episode'] = transformation['episode']
-
-            exists = False
-            try:
-                exists = server.exists(title, item['season'], item['episode'], args = args)
-                uravo.event({'AlertGroup':'server_title', 'AlertKey':title, 'Severity':'green', 'Summary':f"Got {server.name} title for '{title}'"})
-            except Exception as e:
-                if (args.verbose): print(f" ... FAILED: can't get {server.name} title for '{title}'")
-                uravo.event({'AlertGroup':'server_title', 'AlertKey':title, 'Severity':'yellow', 'Summary':f"Can't get {server.name} title for '{title}'"})
-                continue
-            
-            if (exists):
-                if (args.verbose): print(f" ... {title} S{item['season']}E{item['episode']} already in {server.name}")
-                continue
-
-            episode_key = f"{title}|S{item['season']}E{item['episode']}"
-            if ('downloads' not in self.cache): self.cache['downloads'] = {}
-            if (episode_key not in self.cache['downloads']):
-                self.cache['downloads'][episode_key] = { 'item':item }
-            elif (args.force is False and 'date' in self.cache['downloads'][episode_key] and self.cache['downloads'][episode_key]['date'] > datetime.now() - timedelta(hours=1)):
-                if (args.verbose): print(f" ... already downloaded within the last hour")
-                continue
-
-            link_url = item['url']
-            if args.debug: print(f"link_url:{link_url}")
-            if (os.path.exists(config['default']['download_dir']) is False):
-                raise Exception(f"{config['default']['download_dir']} does not exist.")
-            torrent_filename = re.sub(" ", ".", item['name']) + ".torrent"
-            if args.verbose: print(f" ... downloading {torrent_filename} to {config['default']['download_dir']}")
-            if (args.dryrun is False):
-                r = requests.get(link_url, stream=True)
-                with open(f"{config['default']['download_dir']}/{torrent_filename}", "wb") as f:
-                   for chunk in r.iter_content(chunk_size=128):
-                      f.write(chunk)
-                self.cache['downloads'][episode_key]['date'] = datetime.now()
+        self.download(self.rss_feed(), args = args)
 
     @abstractmethod
     def rss_feed(self):
